refactor(hogql): split up printer file
